master: split loss fuctions out of functions.
